refactor: 커서 offset 변환 시 Math.toIntExact() 사용으로 안전성 강화

+++ src/main/java/com/meongnyangerang/meongnyangerang/service/AccommodationSearchService.java
@@ -80,7 +80,7 @@ public class AccommodationSearchService {
               .index("accommodation_room")
               .query(finalQuery)
               .size(pageable.getPageSize())
-              .from((int) pageable.getOffset())
+              .from(Math.toIntExact(pageable.getOffset()))
               .sort(so -> so
                   .field(f -> f
                       .field("totalRating")
